backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
dj-database-url==2.1.0
Django==5.2.1
django-cors-headers==4.3.1
django-filter==23.5
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
hiredis==3.2.1
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
pytest==7.4.3
pytest-django==4.7.0
python-dotenv==1.0.0
pytz==2025.2
redis==6.2.0
requests==2.32.3
setuptools==80.8.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uuid==1.30
whitenoise==6.6.0
